# paasta-monitoring-api/src/kr/paasta/monitoring-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Monitoring Dashboard API",
        "contact": {},
        "version": "5.8.0"
    },
    "paths": {
...
        "/api/v1/cp/workload/container/list": {
            "get": {
                "description": "쿠버네티스 워크로드 컨테이너 리스트를 가져온다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CP"
                ],
                "summary": "쿠버네티스 워크로드 컨테이너 리스트 가져오기",
                "parameters": [
                    {
                        "enum": [
                            "deployment",
                            "statefulset",
                            "daemonset"
                        ],
                        "type": "string",
                        "description": "워크로드명을 주입한다.",
                        "name": "workload",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.BasicResponseForm"
                        }
                    }
                }
            }
        },
        "/api/v1/cp/workload/container/log": {
            "get": {
                "description": "쿠버네티스 컨테이너 로그를 가져온다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CP"
                ],
                "summary": "쿠버네티스 컨테이너 로그 가져오기",
                "parameters": [
                    {
                        "type": "string",
                        "example": "kube-system",
                        "description": "네임스페이스명을 주입한다.",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "kube-proxy",
                        "description": "컨테이너명을 주입한다.",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "kube-proxy-v47r9",
                        "description": "파드명을 주입한다.",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.BasicResponseForm"
                        }
                    }
                }
            }
        },
...
        "/api/v1/saas/app/usage": {
            "get": {
                "description": "애플리케이션(핀포인트 에이전트) 통합 사용량 정보를 가져온다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SaaS"
                ],
                "summary": "애플리케이션(핀포인트 에이전트) 통합 사용량 정보 가져오기",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1m",
                        "description": "현재를 기준으로 조회하고 싶은 기간을 주입한다. 단위는 분(m),  시(h),  일(d)  단위를  사용한다.",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.BasicResponseForm"
                        }
                    }
                }
            }
        },
        "/api/v1/saas/app/usage/list": {
            "get": {
                "description": "애플리케이션(핀포인트 에이전트) 개별 사용량 정보를 가져온다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SaaS"
                ],
                "summary": "애플리케이션(핀포인트 에이전트) 개별 사용량 정보 가져오기",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1m",
                        "description": "현재를 기준으로 조회하고 싶은 기간을 주입한다. 단위는 분(m),  시(h),  일(d)  단위를  사용한다.",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.BasicResponseForm"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiHelpers.BasicResponseForm": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "responseInfo": {},
                "responseTime": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "statusText": {
                    "type": "string"
                }
            }
        },
...
    },
...
}